% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_sto_causal.R
\encoding{UTF-8}
\name{autoplot.est_causal}
\alias{autoplot.est_causal}
\title{Plot an estimated IRF}
\usage{
\method{autoplot}{est_causal}(object, ...)
}
\arguments{
\item{object}{an \code{est_causal} object generated by \code{\link[causirf:est_irf]{causirf::est_irf}}}

\item{...}{other arguments passed to specific methods}
}
\value{
a ggplot object
}
\description{
Plot an estimated IRF
}
\examples{
\dontrun{
library(causirf)

data(temp_co2_data)

obj_causal_temp_co2 <- est_causal(temp_co2_data, "temp", "co2", J = 20, lambda = 10)

autoplot(obj_causal_temp_co2)

evr(obj_causal_co2_temp)
}
}
\seealso{
\code{\link[causirf:est_irf]{causirf::est_irf}},
\code{\link[causirf:evr]{causirf::evr}}
}
