% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causalirf-package.R
\docType{package}
\name{causalirf-package}
\alias{causalirf-package}
\alias{_PACKAGE}
\alias{causalirf}
\title{causalirf}
\description{
Based upon the existence of an impulse response function (IRF)
connecting two stochastic processes, the utils in this package allow to identify
necessary conditions for the existence of a (linear) causal link either from
x to y or y to x, or of a hen-or-egg (HOE) situation in which causal influences
appear to go in both directions.
}
\details{
For more details see the url:

\itemize{
  \item https://github.com/miguel-conde/causirf
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/miguel-conde/causirf}
}

}
