% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_sto_causal.R
\encoding{UTF-8}
\name{cost_fun}
\alias{cost_fun}
\title{Cost function to estimate IRF using least-squares}
\usage{
cost_fun(g, L, J, X, Y, psi, lambda)
}
\arguments{
\item{g}{numeric vector for the IRF, with \code{J} elements}

\item{L, J, X, Y, psi, lambda}{objects used to compute de cost function value}
}
\value{
the value of the cost function at \code{g}
}
\description{
Cost function to estimate IRF using least-squares
}
\keyword{internal}
