% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_sto_causal.R
\encoding{UTF-8}
\name{est_irf}
\alias{est_irf}
\title{Estimate the Impulse Response Function (IRF)}
\usage{
est_irf(in_data, cause, effect, J = 20, lambda = 10)
}
\arguments{
\item{in_data}{a tibble, one of the colummns must account for the cause and
the other for the effect.}

\item{cause}{name of the column accounting for the cause}

\item{effect}{name of the column accounting for the effect}

\item{J}{number of positive and negative lags composing the IRF}

\item{lambda}{weight multiplier to constrain the roughness index (weight of
the penalty from the roughness index)}
}
\value{
an \code{est_causal} object, a list containing:

\itemize{
\item \code{X}, \code{Y}, \code{L}, \code{J}, \code{psi}: objects used in the optimization
\item \code{cause}, \code{effect}: names of the cause and the effect
\item \code{res_optim}: result of \link[stats:optim]{stats::optim}
}
}
\description{
This function estimates the IRF between 2 stochastic processes, one taken
as one of the causes of the other.
}
\examples{
\dontrun{
library(causirf)

data(temp_co2_data)

obj_causal_temp_co2 <- est_causal(temp_co2_data, "temp", "co2", J = 20, lambda = 10)

autoplot(obj_causal_temp_co2)

evr(obj_causal_co2_temp)
}
}
\references{
\itemize{
\item \href{https://royalsocietypublishing.org/doi/10.1098/rspa.2021.0835}{Revisiting causality using stochastics: 1. Theory},
\item \href{https://royalsocietypublishing.org/doi/10.1098/rspa.2021.0836}{Revisiting causality using stochastics: 2. Applications}
}
}
\seealso{
\code{\link[causirf:autoplot.est_causal]{causirf::autoplot}},
\code{\link[causirf:evr]{causirf::evr}}
}
