% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_sto_causal.R
\encoding{UTF-8}
\name{make_probe_x_y}
\alias{make_probe_x_y}
\title{Extends the data with lags}
\usage{
make_probe_x_y(in_data, cause, effect, J = 10)
}
\arguments{
\item{cause}{name of the column accounting for the cause}

\item{effect}{name of the column accounting for the effect}

\item{J}{number of positive and negative lags composing the IRF}

\item{in_dataa}{tibble, one of the colummns must account for the cause and
the other for the effect.}
}
\value{
the \code{in_data} tibble extended with \code{2J+1} lag columns
}
\description{
Extends the data with lags
}
\keyword{internal}
