% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_sto_causal.R
\encoding{UTF-8}
\name{make_psi}
\alias{make_psi}
\title{Build the matrix for the roughness index}
\usage{
make_psi(J)
}
\arguments{
\item{J}{}
}
\value{
a matrix with \code{2J−1} rows and \code{2J+1} columns
}
\description{
Build the matrix for the roughness index
}
\keyword{internal}
